[ricos-content] refactor(preview) exported from /lib
