+++ packages/ricos-content/web/lib/preview.js
@@ -1,0 +1,2 @@
+export { default as ContentStateTransformation } from '../src/preview/RuleEngine/ContentStateTransformation';
+export { INTERACTIONS } from '../src/preview/const';
