+++ packages/ricos-content/web/src/index.ts
@@ -15,5 +15,3 @@ export * from './version';
 export { isContentStateEmpty } from './contentStateUtils/contentStateUtils';
 
 export { createContent } from './contentStateUtils/createContent';
-export { default as ContentStateTransformation } from './preview/RuleEngine/ContentStateTransformation';
-export { INTERACTIONS } from './preview/const';
